+++ src/card.ts
@@ -32,7 +32,6 @@ import './editor/editor';
 import {
   BRAND_LOGO_IMAGE_BASE64,
   BRAND_LOGO_IMAGE_SIZE,
-  FOOTER_ICON_SIZE_DEFAULT,
   PLAYER_CONTROLS_ICON_TOGGLE_COLOR_DEFAULT,
   CARD_CONTENT_MARGIN_DEFAULT
 } from './constants';
@@ -326,7 +325,7 @@ export class Card extends AlertUpdatesBase {
         justify-content: space-around;
         flex-wrap: wrap;
         width: 100%;
-        --mdc-icon-button-size: var(--stpc-footer-icon-button-size, 2.5rem);
+        --ha-icon-button-size: var(--stpc-footer-icon-button-size, 2.5rem);
         --mdc-icon-size: var(--stpc-footer-icon-size, 1.75rem);
         --mdc-ripple-top: 0px;
         --mdc-ripple-left: 0px;
@@ -1210,7 +1209,7 @@ export class Card extends AlertUpdatesBase {
       styleInfo['--stpc-footer-icon-color-selected'] = `${footerIconColorSelected}`;
     if (footerIconSize) {
       styleInfo['--stpc-footer-icon-size'] = `${footerIconSize}`;
-      styleInfo['--stpc-footer-icon-button-size'] = `var(--stpc-footer-icon-size, ${FOOTER_ICON_SIZE_DEFAULT}) + 0.75rem`;
+      styleInfo['--stpc-footer-icon-button-size'] = `var(--stpc-footer-icon-size, ${footerIconSize})`;
     }
     if (footerBackgroundImage)
       styleInfo['--stpc-footer-background-image'] = `${footerBackgroundImage}`;