+++ src/components/player-volume.ts
@@ -9,6 +9,7 @@ import {
 
 // our imports.
 import {
+  PLAYER_CONTROLS_ICON_BUTTON_SIZE_DEFAULT,
   PLAYER_CONTROLS_ICON_SIZE_DEFAULT,
   PLAYER_CONTROLS_ICON_TOGGLE_COLOR_DEFAULT
 } from '../constants';
@@ -334,15 +335,17 @@ class Volume extends AlertUpdatesBase {
     return css`
       ha-control-slider {
         --control-slider-color: var(--stpc-player-volume-slider-color, var(--stpc-player-controls-color, var(--dark-primary-color, ${unsafeCSS(PLAYER_CONTROLS_ICON_TOGGLE_COLOR_DEFAULT)})));
-        --control-slider-thickness: 1rem;
+        --control-slider-thickness: var(--stpc-player-volume-slider-thickness, 1rem);
         border: 1px solid rgba(255, 255, 255, 0.10);
+        border-radius: var(--control-slider-border-radius);
         box-sizing: border-box;
       }
 
       ha-control-slider[disabled] {
         --control-slider-color: var(--disabled-text-color);
-        --control-slider-thickness: 1rem;
+        --control-slider-thickness: var(--stpc-player-volume-slider-thickness, 1rem);
         border: 1px solid rgba(255, 255, 255, 0.10);
+        border-radius: var(--control-slider-border-radius);
         box-sizing: border-box;
       }
 
@@ -381,8 +384,8 @@ class Volume extends AlertUpdatesBase {
       }
 
       *[slim] * {
-        --control-slider-thickness: 10px;
-        --mdc-icon-button-size: 30px;
+        --control-slider-thickness: var(--stpc-player-volume-slider-thickness, 10px);
+        --ha-icon-button-size: 30px;
         --mdc-icon-size: 20px;
       }
 
@@ -403,7 +406,7 @@ class Volume extends AlertUpdatesBase {
         overflow: hidden;
         color: var(--stpc-player-controls-icon-color, #ffffff);
         width: 100%;
-        --mdc-icon-button-size: var(--stpc-player-controls-icon-button-size, 2.75rem);
+        --ha-icon-button-size: var(--stpc-player-controls-icon-button-size, ${unsafeCSS(PLAYER_CONTROLS_ICON_BUTTON_SIZE_DEFAULT)});
         --mdc-icon-size: var(--stpc-player-controls-icon-size, ${unsafeCSS(PLAYER_CONTROLS_ICON_SIZE_DEFAULT)});
       }